Added match amount function in contribution.sol

+++ blockchain/contracts/CrowdFunding.sol
@@ -1,91 +1,105 @@
-pragma solidity >=0.5.0 < 0.9.0;
+pragma solidity >=0.5.0 <0.9.0;
+
 // SPDX-License-Identifier: MIT
 
 // crowd funding smart contract for reference only
-contract crowdFunding{
-    mapping(address=>uint) public contributors;
+contract crowdFunding {
+    mapping(address => uint256) public contributors;
     address public manager;
-    uint public minContribution;
-    uint public deadline;
-    uint public target;
-    uint public raisedAmount;
-    uint public noOfContributors;
+    uint256 public minContribution;
+    uint256 public deadline;
+    uint256 public target;
+    uint256 public raisedAmount;
+    uint256 public noOfContributors;
 
-    struct Request{
+    struct Request {
         string description;
         address payable recipient;
-        uint value;
+        uint256 value;
         bool completed;
-        uint noOfVoters;
-        mapping(address=>bool) voters;
+        uint256 noOfVoters;
+        mapping(address => bool) voters;
     }
 
-    mapping(uint=>Request) public requests;
-    uint public numRequests;
+    mapping(uint256 => Request) public requests;
+    uint256 public numRequests;
 
-    constructor(uint _target, uint _deadline){
+    constructor(uint256 _target, uint256 _deadline) {
         target = _target;
         deadline = block.timestamp + _deadline;
         minContribution = 100 wei;
         manager = msg.sender;
     }
 
-    function sendEther() public payable{
+    function sendEther() public payable {
         require(block.timestamp < deadline, "Deadline has passed");
         require(msg.value > minContribution, "Minimum Contribution is not met");
 
         // if contributor is contributing for the first time
-        if(contributors[msg.sender]==0){
+        if (contributors[msg.sender] == 0) {
             noOfContributors++;
         }
-        contributors[msg.sender]+=msg.value;
-        raisedAmount+=msg.value;
-
+        contributors[msg.sender] += msg.value;
+        raisedAmount += msg.value;
     }
 
-    function getContractBalance() public view returns(uint){
+    function getContractBalance() public view returns (uint256) {
         return address(this).balance;
     }
 
-    function refund() public{
-        require(block.timestamp>deadline && raisedAmount<target, "You are not eligible for refund");
-        require(contributors[msg.sender]>0);
-        address payable user=payable(msg.sender);
+    function refund() public {
+        require(
+            block.timestamp > deadline && raisedAmount < target,
+            "You are not eligible for refund"
+        );
+        require(contributors[msg.sender] > 0);
+        address payable user = payable(msg.sender);
         user.transfer(contributors[msg.sender]);
-        contributors[msg.sender]=0;
+        contributors[msg.sender] = 0;
     }
 
-    modifier onlyManager(){
-        require(msg.sender==manager, "Only manager can call this function");
+    modifier onlyManager() {
+        require(msg.sender == manager, "Only manager can call this function");
         _;
     }
 
-    function createRequest(string memory _description, address payable _recipient, uint _value) public onlyManager{
+    function createRequest(
+        string memory _description,
+        address payable _recipient,
+        uint256 _value
+    ) public onlyManager {
         Request storage newRequest = requests[numRequests];
         numRequests++;
-        newRequest.description=_description;
-        newRequest.recipient=_recipient;
-        newRequest.value=_value;
-        newRequest.completed=false;
-        newRequest.noOfVoters=0;
+        newRequest.description = _description;
+        newRequest.recipient = _recipient;
+        newRequest.value = _value;
+        newRequest.completed = false;
+        newRequest.noOfVoters = 0;
     }
 
-    function voteRequests(uint _requestsNo) public{
-        require(contributors[msg.sender]>0, "you must be a contributor");
+    function voteRequests(uint256 _requestsNo) public {
+        require(contributors[msg.sender] > 0, "you must be a contributor");
         Request storage thisRequest = requests[_requestsNo];
-        require(thisRequest.voters[msg.sender]==false, "you have already voted");
-        thisRequest.voters[msg.sender]=true;
+        require(
+            thisRequest.voters[msg.sender] == false,
+            "you have already voted"
+        );
+        thisRequest.voters[msg.sender] = true;
         thisRequest.noOfVoters++;
     }
 
-    function makePayment(uint _requestsNo) public onlyManager{
-        require(raisedAmount>=target);
+    function makePayment(uint256 _requestsNo) public onlyManager {
+        require(raisedAmount >= target);
         Request storage thisRequest = requests[_requestsNo];
-        require(thisRequest.completed==false, "The request has been completed");
-        require(thisRequest.noOfVoters>noOfContributors/2, "Majority does not support");
+        require(
+            thisRequest.completed == false,
+            "The request has been completed"
+        );
+        require(
+            thisRequest.noOfVoters > noOfContributors / 2,
+            "Majority does not support"
+        );
         thisRequest.recipient.transfer(thisRequest.value);
-        thisRequest.completed=true;
-        
-
+        thisRequest.completed = true;
     }
 }
